requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
PyJWT>=2.8.0

# Database
sqlite3

# AI/LLM
ollama>=0.1.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Production monitoring
prometheus-client>=0.17.0
structlog>=23.1.0
psutil>=5.9.0

# Security
cryptography>=41.0.0
passlib>=1.7.4

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Caching (optional)
redis>=4.6.0

# Docker and deployment
docker>=6.1.0

# AWS (for cloud deployment)
boto3>=1.28.0
botocore>=1.31.0

# Logging and monitoring
colorama>=0.4.6

# Type hints
typing-extensions>=4.7.0

# Data validation
marshmallow>=3.20.0

# Testing utilities
factory-boy>=3.3.0
faker>=19.0.0

# Code quality
pre-commit>=3.3.0
isort>=5.12.0

# Performance profiling
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Security scanning
bandit>=1.7.5
safety>=2.3.0
